fix: move aol and wal to log module
