+++ src/storage/log/mod.rs
@@ -1,10 +1,12 @@
-pub mod wal;
 pub mod aol;
+pub mod wal;
 
 use std::collections::HashMap;
 use std::fmt;
-use std::fs::{read_dir, File, OpenOptions};
-use std::io::{self, Read, Seek, Write};
+use std::fs::File;
+use std::fs::{read_dir, OpenOptions};
+use std::io::BufReader;
+use std::io::{self, BufRead, Read, Seek, SeekFrom, Write};
 use std::os::unix::fs::{FileExt, OpenOptionsExt};
 use std::path::{Path, PathBuf};
 use std::sync::PoisonError;
@@ -1270,6 +1272,113 @@ impl fmt::Display for CorruptionError {
 // Implementation of Error trait for CorruptionError
 impl std::error::Error for CorruptionError {}
 
+pub struct MultiSegmentReader {
+    buf: BufReader<File>,      // Buffer for reading from the current segment.
+    segments: Vec<SegmentRef>, // List of segments to read from.
+    cur: usize,                // Index of current segment in segments.
+    off: usize,                // Offset in current segment.
+}
+
+impl MultiSegmentReader {
+    pub(crate) fn new(segments: Vec<SegmentRef>) -> Result<MultiSegmentReader> {
+        if segments.is_empty() {
+            return Err(Error::IO(IOError::new(
+                io::ErrorKind::InvalidInput,
+                "Empty segment list",
+            )));
+        }
+
+        let cur = 0;
+        let off = 0;
+
+        // Open the first segment's file for reading
+        let mut file = File::open(&segments[cur].file_path)?;
+        file.seek(SeekFrom::Start(segments[cur].file_header_offset))?;
+
+        let buf = BufReader::new(file);
+
+        Ok(MultiSegmentReader {
+            buf,
+            segments,
+            cur,
+            off,
+        })
+    }
+}
+
+impl MultiSegmentReader {
+    fn is_eof(&mut self) -> io::Result<bool> {
+        let is_eof = self.buf.fill_buf()?;
+        Ok(is_eof.is_empty())
+    }
+
+    fn read_to_buffer(&mut self, buf: &mut [u8]) -> io::Result<usize> {
+        let bytes_read = self.buf.read(buf)?;
+        self.off += bytes_read;
+
+        // If we read less than the buffer size, we've reached the end of the current segment.
+        if self.off % BLOCK_SIZE != 0 {
+            // Fill the rest of the buffer with zeros.
+            let i = self.fill_with_zeros(buf, bytes_read);
+            self.off += i;
+            return Ok(bytes_read + i);
+        }
+
+        Ok(bytes_read)
+    }
+
+    fn fill_with_zeros(&mut self, buf: &mut [u8], bytes_read: usize) -> usize {
+        let mut i = 0;
+        while bytes_read + i < buf.len() && (self.off + i) % BLOCK_SIZE != 0 {
+            buf[bytes_read + i] = 0;
+            i += 1;
+        }
+        i
+    }
+
+    fn load_next_segment(&mut self) -> io::Result<()> {
+        if self.cur + 1 >= self.segments.len() {
+            return Err(io::ErrorKind::UnexpectedEof.into());
+        }
+
+        self.cur += 1;
+        self.off = 0;
+
+        let next_file = File::open(&self.segments[self.cur].file_path)?;
+        let header_offset = self.segments[self.cur].file_header_offset;
+        let mut next_buf_reader = BufReader::new(next_file);
+        next_buf_reader.seek(SeekFrom::Start(header_offset))?;
+
+        self.buf = next_buf_reader;
+
+        Ok(())
+    }
+
+    fn current_segment_id(&self) -> u64 {
+        self.segments[self.cur].id
+    }
+
+    fn current_offset(&self) -> usize {
+        self.off
+    }
+}
+
+impl Read for MultiSegmentReader {
+    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
+        if self.cur >= self.segments.len() {
+            return Ok(0);
+        }
+
+        // TODO: could create a problem when reading a partial block spread over multiple segments
+        if !self.is_eof()? {
+            self.read_to_buffer(buf)
+        } else {
+            self.load_next_segment()?;
+            self.read_to_buffer(buf)
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1938,5 +2047,258 @@ mod tests {
 
         // Close the segment
         segment.close().expect("should close segment");
+    }
+
+    fn create_test_segment_ref(segment: &Segment) -> SegmentRef {
+        SegmentRef {
+            file_path: segment.file_path.clone(),
+            file_header_offset: segment.file_header_offset,
+            id: segment.id,
+        }
+    }
+
+    fn create_test_segment(temp_dir: &TempDir, id: u64, data: &[u8]) -> Segment {
+        let opts = Options::default().with_wal();
+        let mut segment = Segment::open(temp_dir.path(), id, &opts).expect("should create segment");
+        let r = segment.append(data);
+        assert!(r.is_ok());
+        assert_eq!(data.len(), r.unwrap().1);
+        segment
+    }
+
+    #[test]
+    fn test_single_segment() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create a sample segment file and populate it with data
+        let mut segment = create_test_segment(&temp_dir, 0, &[0, 1, 2, 3]);
+
+        // Test appending another buffer
+        let r = segment.append(&[4, 5, 6, 7]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Create a Vec of segments containing our sample segment
+        let segments: Vec<SegmentRef> = vec![create_test_segment_ref(&segment)];
+
+        // Create a MultiSegmentReader for testing
+        let mut buf_reader = MultiSegmentReader::new(segments).expect("should create");
+
+        // Read first record from the MultiSegmentReader
+        let mut bs = [0u8; 11];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 11);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        // Read second record from the MultiSegmentReader
+        let mut bs = [0u8; 11];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 11);
+        assert_eq!(&[4, 5, 6, 7].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        let mut bs = [0u8; 11];
+        buf_reader.read(&mut bs).expect_err("should not read");
+
+        assert!(segment.close().is_ok());
+    }
+
+    #[test]
+    fn test_multi_segment() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create a sample segment file and populate it with data
+        let opts = Options::default().with_wal();
+        let mut segment1 = Segment::open(temp_dir.path(), 0, &opts).expect("should create segment");
+        let mut segment2 = Segment::open(temp_dir.path(), 1, &opts).expect("should create segment");
+
+        // Test appending a non-empty buffer
+        let r = segment1.append(&[0, 1, 2, 3]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Test appending another buffer
+        let r = segment2.append(&[4, 5, 6, 7]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Create a Vec of segments containing our sample segment
+        let segments: Vec<SegmentRef> = vec![
+            create_test_segment_ref(&segment1),
+            create_test_segment_ref(&segment2),
+        ];
+
+        // Create a MultiSegmentReader for testing
+        let mut buf_reader = MultiSegmentReader::new(segments).expect("should create");
+
+        // Read first record from the MultiSegmentReader
+        let mut bs = [0u8; 11];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 11);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        // Read second record from the MultiSegmentReader
+        let mut bs = [0u8; 11];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 11);
+        assert_eq!(&[4, 5, 6, 7].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        let mut bs = [0u8; 11];
+        buf_reader.read(&mut bs).expect_err("should not read");
+
+        assert!(segment1.close().is_ok());
+        assert!(segment2.close().is_ok());
+    }
+
+    #[test]
+    fn test_partial_block() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create a sample segment file and populate it with data
+        let opts = Options::default().with_wal();
+        let mut segment = Segment::open(temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test appending a non-empty buffer
+        let r = segment.append(&[1, 2, 3, 4]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Create a Vec of segments containing our sample segment
+        let segments: Vec<SegmentRef> = vec![create_test_segment_ref(&segment)];
+
+        // Create a MultiSegmentReader for testing
+        let mut buf_reader = MultiSegmentReader::new(segments).expect("should create");
+
+        // Read data from the MultiSegmentReader
+        let mut bs = [0u8; 50];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 50);
+        assert_eq!(&[1, 2, 3, 4].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..11]);
+        assert_eq!(buf_reader.off, 50);
+
+        let mut read_buffer = [0u8; 50];
+        buf_reader
+            .read(&mut read_buffer)
+            .expect_err("should not read");
+        assert_eq!(buf_reader.off, 50);
+
+        // sync the segment, and read again
+        assert!(segment.sync().is_ok());
+        let mut read_buffer = [0u8; 50];
+        let bytes_read = buf_reader.read(&mut read_buffer).expect("should read");
+        assert_eq!(bytes_read, 50);
+        assert_eq!(buf_reader.off, 100);
+
+        assert!(segment.close().is_ok());
+    }
+
+    #[test]
+    fn test_full_synced_block() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create a sample segment file and populate it with data
+        let opts = Options::default().with_wal();
+        let mut segment = Segment::open(temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test appending a non-empty buffer
+        let r = segment.append(&[1, 2, 3, 4]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        assert!(segment.sync().is_ok());
+
+        // Create a Vec of segments containing our sample segment
+        let segments: Vec<SegmentRef> = vec![create_test_segment_ref(&segment)];
+        // Create a MultiSegmentReader for testing
+        let mut buf_reader = MultiSegmentReader::new(segments).expect("should create");
+
+        // Read data from the MultiSegmentReader
+        let mut bs = [0u8; 50];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, 50);
+        assert_eq!(&[1, 2, 3, 4].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..11]);
+        assert_eq!(buf_reader.off, 50);
+
+        let mut read_buffer = [0u8; 50];
+        let bytes_read = buf_reader.read(&mut read_buffer).expect("should read");
+        assert_eq!(bytes_read, 50);
+        assert_eq!(buf_reader.off, 100);
+        assert!(read_buffer.iter().all(|&byte| byte == 0));
+
+        assert!(segment.close().is_ok());
+    }
+
+    #[test]
+    fn test_multi_segment_with_sync() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create a sample segment file and populate it with data
+        let opts = Options::default().with_wal();
+        let mut segment1 = Segment::open(temp_dir.path(), 0, &opts).expect("should create segment");
+        let mut segment2 = Segment::open(temp_dir.path(), 1, &opts).expect("should create segment");
+
+        // Test appending a non-empty buffer
+        let r = segment1.append(&[0, 1, 2, 3]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+        assert!(segment1.sync().is_ok());
+
+        // Test appending another buffer
+        let r = segment2.append(&[4, 5, 6, 7]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+        assert!(segment2.sync().is_ok());
+
+        // Create a Vec of segments containing our sample segment
+        let segments: Vec<SegmentRef> = vec![
+            create_test_segment_ref(&segment1),
+            create_test_segment_ref(&segment2),
+        ];
+
+        // Create a MultiSegmentReader for testing
+        let mut buf_reader = MultiSegmentReader::new(segments).expect("should create");
+
+        // Read first record from the MultiSegmentReader
+        let mut bs = [0u8; BLOCK_SIZE];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, BLOCK_SIZE);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..11]);
+
+        // Read second record from the MultiSegmentReader
+        let mut bs = [0u8; BLOCK_SIZE];
+        let bytes_read = buf_reader.read(&mut bs).expect("should read");
+        assert_eq!(bytes_read, BLOCK_SIZE);
+        assert_eq!(&[4, 5, 6, 7].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..11]);
+
+        let mut bs = [0u8; 11];
+        buf_reader.read(&mut bs).expect_err("should not read");
+
+        assert!(segment1.close().is_ok());
+        assert!(segment2.close().is_ok());
+    }
+
+    #[test]
+    fn test_segment_ref() {
+        // Create a temporary directory to hold the segment files
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create sample segment files and populate it with data
+        let mut segment1 = create_test_segment(&temp_dir, 4, &[1, 2, 3, 4]);
+        let mut segment2 = create_test_segment(&temp_dir, 6, &[5, 6]);
+        let mut segment3 = create_test_segment(&temp_dir, 8, &[7, 8, 9]);
+        assert!(segment1.close().is_ok());
+        assert!(segment2.close().is_ok());
+        assert!(segment3.close().is_ok());
+
+        let sr = SegmentRef::read_segments_from_directory(temp_dir.path())
+            .expect("should read segments");
+        assert!(sr.len() == 3);
+        assert!(sr[0].id == 4);
+        assert!(sr[1].id == 6);
+        assert!(sr[2].id == 8);
     }
 }
